Added rudimentary colour to the console renderer.
When I say "rudimentary", I mean it. It's literally just some added ANSI
escape codes to set the foreground colour of the text. Don't even know
if it works in windows.

In all likelihood this branch will never be merged in; it's just to make
things prettier for me while developing.