Generalised pools and added modifiers to Entity.
Pools (HP, MP, and Tech) can now be accessed through a unified
interface, made friendly with templates. This is true for both current
pools and maximum pools.

Also provided `drain' and `restore' functions for Entity. The way it
works is it's impossible to `drain' negative health, as that would be
healing, and vice-versa for `restore'. May need to revisit this design
decision sometime in the future, though.

The reason for the `Entity&' parameter in the pool-modifying functions
is to facilitate reflex actions and observers (for example, for tech
gain). May change my mind later and make it `const Entity&' instead,
though, and queue reflex actions for later instead.