Removed untargetable skills; skills now work (?)
The way it's set up at the moment there's really no room for
"Field"-spread abilities. That'll have to wait until some kind of
battlefield-level modifiers are introduced, I think.

Skills now get passed with the team. I'm not totally happy with how this
is done at the moment, though, so that may change in the future. Teams
themselves are really second-class citizens as it stands; I need to
think more about where they are kept and how they are accessed.