Added descriptors for a skill's power and accuracy.
Also changed the `is_hook' trait to fail on classes attempting to
inherit from multiple different hooks, as that would seriously break
some of the code in weird and wonderful ways.

Descriptor values are calculated from relevant hooks. Currently the
logic is a little obtrusive to the hook creation code, but I couldn't be
bothered to subdivide the function — there would have been almost as
much boilerplate as actual code, at least for the way I initially tried.
I might come back to it later and fix that.

It has also come to my attention that the whole power thing is
completely unbalanced at the moment. As it stands I don't really care
since I'm more worried about getting out functionality then getting a
fun game, but that's something that'll have to change at some point.