Skill action no longer requires non-const Entity
This carried-through to make the renderers need modification access to
ever entity in the battle, which was just weird and wrong. As a side
benefit now the team doesn't need to be passed through the action
either.

Also added a getter for the BattleSystem to determine the team an entity
is on. I might get rid of it, though, as it's nothing that shouldn't be
possible using the current API of BattleSystem anyway.