Added a message logger.
This should hopefully provide a centralised for actions and their
results to be reported in a renderer-agnostic way. It may still need
some fiddling in terms of exactly what messages are returned and the way
that individual messages are clumped together, but that can come after
experimenting a little more.