Fixed bug with removing hooks.
I misunderstood `std::remove_if`; it actually *does* invalidate the
items it removes (it probably moves them, I guess), which means I can't
remove_if, do some work on the things to be removed, and then erase them
safely. Instead I now use std::partition (with a negative condition) to
separate first, and then process and remove.

In actuality I use std::stable_partition to keep the order of hooks the
same. I don't know if that'd ever matter (hopefully it won't, but still
need to think about this), but doing this to stay on the safe side.