Settled on a layout for defining skills in Lua.
I spent an awfully long time just thinking about this, and I've almost
come to the point where I could conceivably take what I've learnt and
apply it to my non-Lua code. Still, this is useful knowledge to have, so
I'll keep pushing it.

This is very much not complete, yet, but it should work for most general
things. Perks, buffs, and documentation are still to come, as well as
full implementations of some of the things going on here that are
currently left as mere placeholders.