Moved skill details into a config file.
This is paving the way to include Entity configuration in the game, as
well. Also just useful for cleaning up the code.

At some point config.cpp should probably be split up, possibly. We'll
see. Also will need to redesign the ways skills are represented: they
can be much simpler now, since SkillDetails is (for all intents and
purposes) immutable -- or at least, it should be.