Moved creation of entity details from entity class
This way it's the entity creator's responsibility to set up the name,
stats, level, and so forth for an entity. Gives us more flexibility for
entities that are created in different ways, and is a fallback for now
as I decide exactly how I want to define entities in config files.

Also fixed some things up with how controllers are applied, and removed
some stuff from Skill I really should have removed ages back.