Fixed bug getting teams by passing in BattleSystem
Originally you couldn't (easily) tell what team the target was in, and
the `getTeam` function would just always return the enemy team for the
target whether or not the target's team _was_ actually the enemy team.

To fix this we now pass in the whole BattleSystem running the show. It
can be used to retrieve the team for a given entity. It's also
eventually useful as we can assign environmental effects in the system,
as well as allow functionality to add entities dynamically to the
battle.

TODO, though, is make BattleSystem better OOP (encapsulate, single
responsibility) and propagate those changes to these systems.