Cleaned up lua, added global 'log' function
This is mostly focussed on improvements for the lua side of things,
though there is a little cleaning up on the C++ side as well. The
benefits for this commit come from the 'luacheck' tool, which I should
set up a config file for (once I work out how to do that!).

`func.lua` should eventually be filled with common functionality between
different kinds of skills. Eventually I'll need proper documentation to
deal with it as well.