Changed calculation of turn order.
Renamed `speed' to `react' to better capture the new semantics. In
essence, what we do is just say "higher `react' = can move through the
battle faster, and thus gets more turns on average."

Something to think about in the future, though, is to maybe use
std::make_heap instead of a std::priority_queue, to allow us to easily
remove or change arbritrary elements in the turn order. It depends on
what semantics we decide for that, of course.